backend/requirements.txt
fastapi
uvicorn
python-multipart
Orange3
